[UPDATE] schedule bar display algorithm
